Allow completion styles to have sorting functions, too
Add a sorting function to the 'flex' completion style that sorts
according to completion score.

When the source of the completion strings -- the completion table --
doesn't specify a sorting function in its 'metadata' alist (be it a
'cycle-sort-function' or a 'display-sort-function') we ask the
completion style for a sorting solution.  Only in its absence do we
fallback to default sorting, as before.

This is especially useful for completion styles such as 'flex', which
score each matched candidated according to the likelihood that the
match was intended by the search pattern.  The new
'completion-flex-sort-by-score', which considers these pre-computed
scores, is much better than the default strategy which is normally
counter-productive in these cases.

* lisp/minibuffer.el (completion--nth-completion): Store used
style in metadata.
(completion-all-sorted-completions, minibuffer-completion-help):
Consider the style's sorting function, if it exists.
(flex): Give this completion style a sorting function.
(completion-flex-sort-by-score): Sort by pre-computed score.