Eglot: symbol highlighting now done in eldoc-documentation-functions
Even though highlighting the thing at point isn't really
producing documentation at point, theory and practice have shown
that using eldoc-documentation-functions for this is very
feasible and a good way to avoid reimplementing all the idle
timer logic for such functionality.

We were already using ElDoc for this purpose, but this commit
makes it so the new eglot-highlight-eldoc-function is a first
class citizen in eldoc-documentation-functions, so users can
manipulate it if so desired.

* lisp/progmodes/eglot.el (eglot--managed-mode): Use
eglot-highlight-eldoc-function
(eglot-hover-eldoc-function): Don't mix highlighting with hover.
(eglot-highlight-eldoc-function): Rename and rework from
eglot--highlight-piggyback